+++ src/sql/src/plan.rs
@@ -36,8 +36,7 @@ use dataflow_types::{SinkConnectorBuilder, SourceConnector};
 use repr::{ColumnName, RelationDesc, Row, ScalarType, Timestamp};
 use sql_parser::ast::Expr;
 
-use crate::ast::{ExplainOptions, ExplainStage, ObjectType, Statement};
-use crate::catalog::Catalog;
+use crate::ast::{ExplainOptions, ExplainStage, FetchDirection, ObjectType, Statement};
 use crate::names::{DatabaseSpecifier, FullName, SchemaName};
 
 pub(crate) mod decorrelate;
@@ -45,6 +44,7 @@ pub(crate) mod error;
 pub(crate) mod explain;
 pub(crate) mod expr;
 pub(crate) mod func;
+pub(crate) mod plan_utils;
 pub(crate) mod query;
 pub(crate) mod scope;
 pub(crate) mod statement;
@@ -57,7 +57,7 @@ pub use error::PlanError;
 pub use explain::Explanation;
 // This is used by sqllogictest to turn SQL values into `Datum`s.
 pub use query::{scalar_type_from_sql, unwrap_numeric_typ_mod};
-pub use statement::{describe_statement, FetchOptions, StatementContext, StatementDesc};
+pub use statement::{describe, plan, StatementContext, StatementDesc};
 
 /// Instructions for executing a SQL query.
 #[derive(Debug)]
@@ -167,6 +167,18 @@ pub enum Plan {
         object_type: ObjectType,
     },
     AlterIndexLogicalCompactionWindow(Option<AlterIndexLogicalCompactionWindow>),
+    Declare {
+        name: String,
+        stmt: Statement,
+    },
+    Fetch {
+        name: String,
+        count: Option<FetchDirection>,
+        timeout: ExecuteTimeout,
+    },
+    Close {
+        name: String,
+    },
 }
 
 #[derive(Clone, Debug)]
@@ -246,6 +258,13 @@ pub enum CopyFormat {
     Binary,
 }
 
+#[derive(Debug, Copy, Clone)]
+pub enum ExecuteTimeout {
+    None,
+    Seconds(f64),
+    WaitOnce,
+}
+
 #[derive(Debug, PartialEq)]
 pub struct AlterIndexLogicalCompactionWindow {
     pub index: GlobalId,
@@ -291,29 +310,4 @@ impl Default for PlanContext {
             wall_time: Utc::now(),
         }
     }
-}
-
-/// Produces a [`Plan`] from the purified statement `stmt`.
-///
-/// Planning is a pure, synchronous function and so requires that the provided
-/// `stmt` does does not depend on any external state. To purify a statement,
-/// use [`crate::pure::purify`].
-pub fn plan(
-    pcx: &PlanContext,
-    catalog: &dyn Catalog,
-    stmt: Statement,
-    params: &Params,
-) -> Result<Plan, anyhow::Error> {
-    statement::handle_statement(pcx, catalog, stmt, params)
-}
-
-/// Creates a description of the purified statement `stmt`.
-///
-/// See the documentation of [`StatementDesc`] for details.
-pub fn describe(
-    catalog: &dyn Catalog,
-    stmt: Statement,
-    param_types: &[Option<pgrepr::Type>],
-) -> Result<StatementDesc, anyhow::Error> {
-    statement::describe_statement(catalog, stmt, param_types)
 }
